Fix for HARMONY-791 ([classlib][lang] unexpected NegativeArraySizeException for java.lang.String(int[], 2, Integer.MAX_VALUE))
git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@420929 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/luni/src/main/java/java/lang/String.java
@@ -498,7 +498,7 @@ public final class String implements Serializable, Comparable<String>, CharSeque
         super();
         if (codePoints == null)
             throw new NullPointerException();
-        if (offset < 0 || count < 0 || (offset + count) > codePoints.length)
+        if (offset < 0 || count < 0 || (long)offset + (long)count > codePoints.length)
             throw new IndexOutOfBoundsException();
         this.offset = 0;
         this.value = new char[count * 2];
